connect client and server

+++ server/src/Post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { Injectable, NotFoundException } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
 import { MyPost } from './schema/post.entity'
@@ -14,6 +14,12 @@ export class PostService {
     async create(createPostDto: CreatePostDto): Promise<MyPost> {
         const { title, body } = createPostDto
 
+        if (!title || !body) {
+            throw new NotFoundException(
+                'Не удалось создать пост: Заголовок и текст не должны быть пустыми!',
+            )
+        }
+
         const newPost = this.postsRepository.create({
             title,
             body,
@@ -23,15 +29,23 @@ export class PostService {
         return savedPost
     }
 
-    async findAll(): Promise<MyPost[]> {
-        return await this.postsRepository.find()
+    async findAll(page: number, limit: number): Promise<MyPost[]> {
+        return await this.postsRepository.find({
+            take: limit,
+            skip: (page - 1) * 12,
+            order: {
+                id: 'DESC',
+            },
+        })
     }
 
     async findById(id: number): Promise<MyPost | null> {
         return this.postsRepository.findOneBy({ id })
     }
 
-    async remove(id: number): Promise<void> {
-        await this.postsRepository.delete(id)
+    async remove(id: number | number[]): Promise<void> {
+        if (id) {
+            await this.postsRepository.delete(id)
+        }
     }
 }
